Simplified service by adding logger in service options
