+++ composer.json
@@ -18,7 +18,8 @@
         "php": ">=5.3.3",
         "guzzle/guzzle": "~3.9",
         "itkg/core": "dev-master",
-        "symfony/options-resolver": "~2"
+        "symfony/options-resolver": "~2",
+        "jms/serializer": "0.*"
     },
     "autoload": {
         "psr-0": {
