+++ src/Itkg/Consumer/Listener/LoggerListener.php
@@ -4,7 +4,6 @@ namespace Itkg\Consumer\Listener;
 
 use Itkg\Consumer\Event\ServiceEvent;
 use Itkg\Consumer\Event\ServiceEvents;
-use Itkg\Consumer\Service\ServiceLoggableInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
 /**
@@ -22,8 +21,8 @@ class LoggerListener implements EventSubscriberInterface
     public function onServiceRequest(ServiceEvent $event)
     {
         $service = $event->getService();
-        if ($service instanceof ServiceLoggableInterface) {
-            $service->getLogger()->info('Request will be send', array(
+        if ($service->getOption('loggable')) {
+            $service->getOption('logger')->info('Request will be send', array(
                 'service' => $service
             ));
         }
@@ -35,8 +34,8 @@ class LoggerListener implements EventSubscriberInterface
     public function onServiceResponse(ServiceEvent $event)
     {
         $service = $event->getService();
-        if ($service instanceof ServiceLoggableInterface) {
-            $service->getLogger()->info('Response success', array(
+        if ($service->getOption('loggable')) {
+            $service->getOption('logger')->info('Response success', array(
                 'service' => $service
             ));
         }
@@ -48,8 +47,8 @@ class LoggerListener implements EventSubscriberInterface
     public function onServiceException(ServiceEvent $event)
     {
         $service = $event->getService();
-        if ($service instanceof ServiceLoggableInterface) {
-            $service->getLogger()->error('Response KO', array(
+        if ($service->getOption('loggable')) {
+            $service->getOption('logger')->error('Response KO', array(
                 'service' => $service
             ));
         }
