--- src/Itkg/Consumer/Service/ServiceLoggableInterface.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace Itkg\Consumer\Service;
-
-use Psr\Log\LoggerInterface;
-
-interface ServiceLoggableInterface
-{
-    /**
-     * @return LoggerInterface
-     */
-    public function getLogger();
-
-    /**
-     * @param LoggerInterface $logger
-     * @return $this
-     */
-    public function setLogger(LoggerInterface $logger);
-}
